+++ src/main/java/com/example/adambackend/config/WebMvcConfig.java
@@ -8,7 +8,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 @Configuration
 public class WebMvcConfig implements WebMvcConfigurer {
 
-        private final long MAX_AGE_SECS = 3600;
+        private final Integer MAX_AGE_SECS = 3600;
 
         @Value("${app.cors.allowedOrigins}")
         private String[] allowedOrigins;
